fix: make timeout anchor recovery durable
